FIX: 各画面の切り分け
