+++ lib/main.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'firebase_options.dart';
+import 'login_screen.dart';
+import 'create_user_screen.dart';
+import 'reset_password_screen.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -28,208 +30,6 @@ class MyApp extends StatelessWidget {
         '/createUser': (context) => const CreateUserScreen(),
         '/resetPassword': (context) => const ResetPasswordScreen(),
       },
-    );
-  }
-}
-
-class LoginScreen extends StatefulWidget {
-  const LoginScreen({super.key});
-
-  @override
-  _LoginScreenState createState() => _LoginScreenState();
-}
-
-class _LoginScreenState extends State<LoginScreen> {
-  String _email = '';
-  String _password = '';
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: const Text('ログイン')),
-      body: Center(
-        child: Padding(
-          padding: const EdgeInsets.all(30),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              TextFormField(
-                decoration: const InputDecoration(labelText: 'メールアドレス'),
-                onChanged: (String value) {
-                  setState(() {
-                    _email = value;
-                  });
-                },
-              ),
-              TextFormField(
-                decoration: const InputDecoration(labelText: 'パスワード'),
-                obscureText: true,
-                onChanged: (String value) {
-                  setState(() {
-                    _password = value;
-                  });
-                },
-              ),
-              Padding(
-                padding: const EdgeInsets.only(top: 40, bottom: 20),
-                child: ElevatedButton(
-                  child: const Text('ログイン'),
-                  onPressed: () async {
-                    try {
-                      final User? user = (await FirebaseAuth.instance.signInWithEmailAndPassword(
-                          email: _email, password: _password))
-                          .user;
-                      if (user != null) {
-                        print("ログインしました ${user.email} , ${user.uid}");
-                      }
-                    } catch (e) {
-                      print(e);
-                    }
-                  },
-                ),
-              ),
-              TextButton(
-                child: const Text('ユーザ登録画面へ'),
-                onPressed: () {
-                  Navigator.pushNamed(context, '/createUser');
-                },
-              ),
-              TextButton(
-                child: const Text('パスワードリセット画面へ'),
-                onPressed: () {
-                  Navigator.pushNamed(context, '/resetPassword');
-                },
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class CreateUserScreen extends StatefulWidget {
-  const CreateUserScreen({super.key});
-
-  @override
-  _CreateUserScreenState createState() => _CreateUserScreenState();
-}
-
-class _CreateUserScreenState extends State<CreateUserScreen> {
-  String _email = '';
-  String _password = '';
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: const Text('ユーザ登録')),
-      body: Center(
-        child: Padding(
-          padding: const EdgeInsets.all(24),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              TextFormField(
-                decoration: const InputDecoration(labelText: 'メールアドレス'),
-                onChanged: (String value) {
-                  setState(() {
-                    _email = value;
-                  });
-                },
-              ),
-              TextFormField(
-                decoration: const InputDecoration(labelText: 'パスワード'),
-                obscureText: true,
-                onChanged: (String value) {
-                  setState(() {
-                    _password = value;
-                  });
-                },
-              ),
-              Padding(
-                padding: const EdgeInsets.only(top: 40, bottom: 20),
-                child: ElevatedButton(
-                  child: const Text('ユーザ登録'),
-                  onPressed: () async {
-                    try {
-                      final User? user = (await FirebaseAuth.instance.createUserWithEmailAndPassword(
-                          email: _email, password: _password))
-                          .user;
-                      if (user != null) {
-                        print("ユーザ登録しました ${user.email} , ${user.uid}");
-                      }
-                    } catch (e) {
-                      print(e);
-                    }
-                  },
-                ),
-              ),
-              TextButton(
-                child: const Text('ログイン画面へ'),
-                onPressed: () {
-                  Navigator.pushNamed(context, '/login');
-                },
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class ResetPasswordScreen extends StatefulWidget {
-  const ResetPasswordScreen({super.key});
-
-  @override
-  _ResetPasswordScreenState createState() => _ResetPasswordScreenState();
-}
-
-class _ResetPasswordScreenState extends State<ResetPasswordScreen> {
-  String _email = '';
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: const Text('パスワードリセット')),
-      body: Center(
-        child: Padding(
-          padding: const EdgeInsets.all(24),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              TextFormField(
-                decoration: const InputDecoration(labelText: 'メールアドレス'),
-                onChanged: (String value) {
-                  setState(() {
-                    _email = value;
-                  });
-                },
-              ),
-              Padding(
-                padding: const EdgeInsets.only(top: 40, bottom: 20),
-                child: ElevatedButton(
-                  child: const Text('パスワードリセット'),
-                  onPressed: () async {
-                    try {
-                      await FirebaseAuth.instance.sendPasswordResetEmail(email: _email);
-                      print("パスワードリセット用のメールを送信しました");
-                    } catch (e) {
-                      print(e);
-                    }
-                  },
-                ),
-              ),
-              TextButton(
-                child: const Text('ログイン画面へ'),
-                onPressed: () {
-                  Navigator.pushNamed(context, '/login');
-                },
-              ),
-            ],
-          ),
-        ),
-      ),
     );
   }
 }
